Move CRU dashboard methods to new controller
We’ll be adding some new functionality to the CRU dashboard soon
that aren’t Placement Request related, so it makes sense to make
these methods agnostic. I’ve also redirected the old paths to the new
paths using our new RedirectController.
