+++ server/views/admin/cruDashboard/search.njk
@@ -25,7 +25,7 @@
 
       <div class="search-and-filter__wrapper">
 
-        <form action="{{ paths.admin.placementRequests.search({}) }}" method="get">
+        <form action="{{ paths.admin.cruDashboard.search({}) }}" method="get">
 
           <input type="hidden" name="_csrf" value="{{ csrfToken }}"/>
